// src/interfaces/Starships.ts
export default interface Starships {
  name: string;
  model: string;
  manufacturer: string;
  cost_in_credits: string;
  length: string;
  max_atmosphering_speed: string;
  crew: string;
  passengers: string;
  cargo_capacity: string;
  consumables: string;
  hyperdrive_rating: string;
  MGLT: string;
  url: string;
  starship_class: string;
  pilots: Array<string>;
  films: Array<string>;
}